Resolve linhas extras lint
